created eventHelper, child of events component and edited events component
