+++ app/components/Home.js
@@ -5,9 +5,6 @@ var React = require("react");
 var Query = require("./search/Query");
 var Results = require("./search/Results");
 
-// Include the helpers for making API calls
-var helpers = require("../utils/helpers");
-
 // Create the Search component
 var Home = React.createClass({
