+++ app/components/Saved.js
@@ -1,9 +1,6 @@
 // Include React as a dependency
 var React = require("react");
 
-// Include the Helper (for the saved recall)
-var helpers = require("../utils/helpers");
-
 // Create the Main component
 var Main = React.createClass({
