+++ app/components/Search.js
@@ -5,8 +5,6 @@ var React = require("react");
 var Query = require("./search/Query");
 var Results = require("./search/Results");
 
-// Include the helpers for making API calls
-var helpers = require("../utils/helpers");
 
 // Create the Search component
 var Search = React.createClass({
